First Round Of Real Code
This is the first actual code thats been made, specifically in regards to using platformio instead of the arduino ide.

Added in the ability to send floats as binary arrays, and multiple of them at that. Another commit will be added later for reading them, but for now they can be sent.