Updated Base Communication And MS5611 Library
Got the base radio code to send data between two radios, and its the full data minus the gps. So this sends the baromter and gyro + callsign.
Also included is a custom "library" for the MS5611. Made it to have full control over spi without having to decrypt how the provided library works